fix(sessions): fix logic for deleting a specific session

+++ src/modules/user-accounts/application/usecases/sessions/delete-session.usecase.ts
@@ -2,6 +2,8 @@ import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { SessionContextDto } from '../../../guards/dto/session-context.dto';
 import { SessionsRepository } from '../../../infrastructure/sessions.repository';
 import { SessionDocument } from 'src/modules/user-accounts/domain/entities/session/session.entity';
+import { DomainException } from '../../../../../core/exceptions/damain-exceptions';
+import { DomainExceptionCode } from '../../../../../core/exceptions/domain-exception-codes';
 
 export class DeleteSessionCommand {
   constructor(public readonly dto: SessionContextDto) {}
@@ -14,11 +16,22 @@ export class DeleteSessionUseCase
   constructor(private readonly sessionsRepository: SessionsRepository) {}
 
   async execute({ dto }: DeleteSessionCommand): Promise<string> {
-    const session: SessionDocument =
-      await this.sessionsRepository.getByUserIdAndDeviceIdOrNotFoundFail(
-        dto.userId,
-        dto.deviceId,
-      );
+    const session: SessionDocument | null =
+      await this.sessionsRepository.getByDeviceId(dto.deviceId);
+
+    if (!session) {
+      throw new DomainException({
+        code: DomainExceptionCode.NotFound,
+        message: `The session with ID (${dto.deviceId}) does not exist`,
+      });
+    }
+
+    if (session.userId !== dto.userId) {
+      throw new DomainException({
+        code: DomainExceptionCode.Forbidden,
+        message: `The user does not have permission to delete this session`,
+      });
+    }
 
     session.delete();
